Attempt 1 at improving the figure
I'd like to have a trend for the training data and for the testing data but can't seem to make it work both at the same time with the function.
So I'm creating it myself but with that many species it's not readable at once.
Other option of adding the facet later but it's crowded at well. Commiting it just so that we have the history for the future if we focus on a couple of species.
We might also decide that we only want one of the two trends.